Add my setup to the sources of the site.
This way, I don't have to carry around the setup file separately from the
source of the whole site. As far as I am concerned, the setup of the site
*is* a source of the site and it should be kept with it, especially for
sites that are statically created (like this one).

I wish that Joey Hess allows this on future versions of ikiwiki.